Preserving full path is optional

+++ git-import.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
 appname=${0##*/}
+debug=1
+
+unset preserve_paths
+if [ "$1" = '-p' ] || [ "$1" = '--preserve-paths' ]; then
+	preserve_paths=1
+	shift
+fi
+
 src_repo=$1
 src_branch=$2
 src_dir=${3#$src_repo}
@@ -8,18 +16,22 @@ src_dir=${3#$src_repo}
 if [ -z "$src_repo" ] || [ -z "$src_branch" ] || [ -z "$src_dir" ]; then
 	echo >&2 "Imports a directory/file in a branch of another repository
 to the current branch of this repository, preserving the history
-and relative path of the file(s).
+and optionally the relative path of the file(s).
 
-Usage: $appname SRC_REPO SRC_BRANCH SRC_DIR [SRC_FILE...]
+Usage: $appname [-p] SRC_REPO SRC_BRANCH SRC_DIR [SRC_FILE...]
+
+-p, --preserve-paths
+              Preserve the paths of source files.  If not specified,
+              the common ancestor of source files is used as the project root.
 
 SRC_REPO      Source repository path or URI
-SRC_BRANCH    Source branch.
+SRC_BRANCH    Source branch
 SRC_DIR       Source directory.  If it can be resolved to a directory in the
                 local filesystem after removing the SRC_REPO prefix, that
                 directory is used.
-SRC_FILE      Source file path relative to SRC_DIR.  If it can be resolved
-                to a file in the local filesystem after processing SRC_DIR
-                as described above, that file is used."
+SRC_FILE      Source file paths relative to SRC_DIR.  If they can be resolved
+                to files in the local filesystem after processing SRC_DIR
+                as described above, those files are used."
 	exit 1
 fi
 
@@ -41,7 +53,7 @@ shift 3
 for src_file in "$@"
 do
 	[ -n "$src_files" ] && src_files="$src_files|"
-	src_file=${src_file#$src_repo}
+	src_file=${src_file#$src_repo/}
 	src_file=${src_file#$src_dir}
 	src_file=${src_file##/}
 	src_files=$src_files${src_file}
@@ -65,20 +77,24 @@ git clone --branch "$src_branch" "$src_repo" "$tmpdir" &&
 				--all
 		fi
 	) &&
-	mkdir -p -- "$src_dir" &&
 	(
-		for file in *
-		do
-			if [ -e "$file" ] && [ "${src_dir#$file}" = "$src_dir" ]; then
-				git mv -- "$file" "$src_dir"
-			fi
-		done
+		if [ $preserve_paths ]; then
+			mkdir -p -- "$src_dir" &&
+				(
+					for file in *
+					do
+						if [ -e "$file" ] && [ "${src_dir#$file}" = "$src_dir" ]; then
+							git mv -- "$file" "$src_dir"
+						fi
+					done
+				) &&
+				git add --all . &&
+				git commit --signoff --message="$appname: Restored filtered files to '$src_dir/'"
+		fi
 	) &&
-	git add --all . &&
-	git commit --signoff --message="$appname: Restored filtered files to '$src_dir/'" &&
 	cd - &&
 	git remote add "$remote" "$tmpdir" &&
 		git fetch "$remote" "$src_branch" &&
 		git merge --edit --message="$appname: Merged '$src_dir/$src_files' from branch '$src_branch' of $repo_name" FETCH_HEAD &&
 	git remote rm "$remote" &&
-	rm -rf "$tmpdir"
+	[ ! $debug ] && rm -rf "$tmpdir"
